finance_API/requirements.txt
asgiref==3.8.1
astroid==3.2.2
colorama==0.4.6
dill==0.3.8
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
flake8==7.1.0
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
tomlkit==0.13.0
tzdata==2024.1
